Add files via upload
In this cutting-edge project, I explore the capabilities of Generative Adversarial Networks (GANs) to create realistic and high-quality images. Leveraging the innovative architecture of GANs, my goal is to develop a model that can generate images indistinguishable from real photographs. This project showcases the potential of GANs in various applications, from art generation to data augmentation.